pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jpresume"
version = "0.1.0"
description = "Convert western-style resumes to Japanese format (履歴書・職務経歴書)"
readme = {text = "Convert western-style resumes to Japanese format", content-type = "text/plain"}
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "mistune>=3.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "openai>=1.0",
    "fpdf2>=2.8",
    "reportlab>=4.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
dev = ["pytest>=8.0", "pytest-mock>=3.0"]

[project.scripts]
jpresume = "jpresume.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jpresume"]

[tool.pytest.ini_options]
testpaths = ["tests"]
